Move DislplayToken to new file
